GitWrap/Errors: Rename Error.h to Result.h
